converting amof into a eclipse plugin. bugfixes merge.

+++ AMOF2/src/hub/sam/mof/ant/MergeModel.java
@@ -82,7 +82,8 @@ public class MergeModel extends Task {
             }
 
             for(cmof.reflection.Object obj: toDelete) {
-                if (((ObjectImpl)obj).getClassInstance().isValid()) {
+            	if (((ObjectImpl)obj).getClassInstance() != null &&
+            			((ObjectImpl)obj).getClassInstance().isValid()) {
                     obj.delete();
                 }
             }
